+++ src/main/ipcs/bs-maps-ipcs.ts
@@ -1,5 +1,5 @@
 import { ipcMain } from "electron";
-import { LocalMapsManagerService } from "../services/maps/local-maps-manager.service";
+import { LocalMapsManagerService } from "../services/additional-content/local-maps-manager.service";
 import { UtilsService } from "../services/utils.service";
 import { BSVersion } from "shared/bs-version.interface";
 import { IpcRequest } from "shared/models/ipc";
