+++ src/main/services/additional-content/local-maps-manager.service.ts
